requirements.txt
flask[async]>=3.0.0
python-dotenv==1.0.0
requests==2.31.0
renault-api==0.1.11
quart==0.19.4